move all backend stuff under app; rename /<date> route to /date/<date>; rename app.py to api.py and app to flak_app in api.py
