+++ app/src/scripts/run_api.sh
@@ -1,0 +1,1 @@
+python app/src/scripts/run_api.py
