--- src/scripts/run_api.sh
@@ -1,1 +1,0 @@
-python src/scripts/run_api.py
